requirements.txt
PyQt6>=6.6.0
mysql-connector-python>=8.3.0
requests>=2.31.0
python-dotenv>=1.0.0
